fix: :bug: reboot if sda disk is wrong
The script normally looks for /dev/sda to create the exroot, the problem with this is that if there are two or more disks, and the first one (which would normally be sda) fails, the second one (which would be sdb) is set to sda and consequently the script will format the wrong disk.  This change tries to exactly deal with that, using an array with the uuids of the disks that should be IGNORED, it checks if any of those disks is the sda, if it is, the script forces the device to restart (and can be changed to just terminate the script).