use blockdev --rereadpt instead of sleep
FTR, this may not be needed. i have realized after this change that
the pendrive i was using was too small and fdisk errored out in a
way that was not visible in the log, only its symptoms.

but this is a nicer way than arbitrary sleeps for not much extra space.